tgloggerd: store Unix timestamps as 64-bit
MySQL INT is 32-bit signed, so a Unix timestamp column overflows after
2038-01-19 (the year 2038 problem). Widen every Unix-timestamp column to
BIGINT: date and edit_date (messages and edits), origin_date (fwd_info),
and users.emoji_status_expiration_date. Non-timestamp INT columns (color
ids, positions) are left as INT.

Widen the pipeline to match so it works past 2038, not just stores:
the model fields become int64_t, and old_edit_date is read back with
std::stoll instead of std::stoi (which would throw on a > 2^31 value).

Note: TDLib currently delivers these as int32, which remains the upstream
32-bit limit until TDLib widens them; the storage and our own code path
are now 64-bit ready.

+++ migrations/000002_create_users_table.up.sql
@@ -28,7 +28,7 @@ CREATE TABLE users (
 
 	-- td_api::emojiStatus
 	emoji_status_custom_emoji_id       BIGINT          NULL COMMENT 'Custom emoji id shown as emoji status; NULL if none.',
-	emoji_status_expiration_date       INT             NULL COMMENT 'Unix time when the emoji status expires; NULL if none.',
+	emoji_status_expiration_date       BIGINT          NULL COMMENT 'Unix time when the emoji status expires; NULL if none.',
 
 	-- td_api::verificationStatus
 	is_verified                        TINYINT(1)      NOT NULL DEFAULT 0 COMMENT 'User is verified by Telegram.',
