+++ src/tgloggerd/DB.hpp
@@ -121,9 +121,9 @@ private:
 					  const char *fk_column,
 					  uint64_t message_row_id,
 					  const models::MessageContent &old_content,
-					  int32_t old_edit_date,
+					  int64_t old_edit_date,
 					  const models::MessageContent &new_content,
-					  int32_t new_edit_date);
+					  int64_t new_edit_date);
 	void insertForwardInfo(mysql::Transaction &tx, const char *table,
 			       const char *fk_column, uint64_t message_row_id,
 			       const models::ForwardInfo &info);
